Inventory System!
Imported a free online inventory system that works really well. Instead
of reinventing the wheel (which would have taken at least 5 hours of
work) we'll use this - it works fantastically. Also makes it very simple
to add items and such. We can replace the graphic UI of it too.
Currently, press 'E' to open.